feat(quant): scale-tensor layout descriptors and params_dtype through the layer stack
- Add kernels/ops/quantization/scale_layout.py: allocate per-token-group
  scale grids directly in the consumer's format (column-major TMA view,
  4-row alignment) so the quantiser writes once and no transpose/pad/cat
  rides the per-layer critical path.
- per_token_group_quant accepts caller-owned output_q/output_s, validates
  shapes and infers the layout from a handed-in buffer's strides; DeepGEMM
  declares the layout it reads and allocates into it.
- Rename LinearBase/parallel-linear ``dtype`` -> ``params_dtype`` (vLLM auto
  convention): the layer prescribes no precision of its own -- models pass
  config.dtype, direct instantiation defers to torch.get_default_dtype().
- Re-export per_token_group_quant from kernels.ops.quantization (a rebase
  had dropped it); refresh the quant matrix doc to the int4byte tables.

+++ lite_llama/modules/mlp.py
@@ -54,10 +54,18 @@ class FusedMLP(nn.Module):
         local_inter = divide(inter, get_tensor_model_parallel_world_size(), "MLP intermediate")
         _check_shard_alignment(quant, local_inter, "MLP intermediate")
         self.gate_up_proj = ColumnParallelLinear(
-            hidden, 2 * inter, quant=quant, dtype=config.dtype, what="MLP intermediate"
+            hidden,
+            2 * inter,
+            quant=quant,
+            params_dtype=config.dtype,
+            what="MLP intermediate",
         )
         self.down_proj = RowParallelLinear(
-            inter, hidden, quant=quant, dtype=config.dtype, what="MLP intermediate"
+            inter,
+            hidden,
+            quant=quant,
+            params_dtype=config.dtype,
+            what="MLP intermediate",
         )
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
